Stable row detection?
At least in b-em and jsbeeb this *seems* to give consistent "colour
bands" across each (text) row of the screen. This is hacky as hell but
it's a good point to stop. (I am moving towards a system where I can
"wait for the raster to pass" if it's coming up behind me when I'm about
to toggle an LED in video RAM, so as to avoid tearing.)