Remove TODO
I've dumped period_table and done some basic histogram plots of period
vs frequency and it does seem reasonable - the upper and lower bounds
are what I'd expect, and the distribution shape is *vaguely* flat or
binomial as appropriate. (I just did this for the default frequency and
spread.) I might have expected the distributions to look more
"stereotypical" but I really have no intuitive idea how large a sample
size is necessary and I'm not too worried about it.